Module fil_actor_market_state::v13

Modules

balance_table

policy

Structs

ActivatedDeal

AddBalanceParams

BatchActivateDealsParams

BatchActivateDealsResult

ClientDealProposal

ClientDealProposal is a DealProposal signed by a client

DealProposal

Note: Deal Collateral is only released and returned to clients and miners when the storage deal stops counting towards power. In the current iteration, it will be released when the sector containing the storage deals expires, even though some storage deals can expire earlier than the sector does. Collaterals are denominated in PerEpoch to incur a cost for self dealing or minimal deals that last for a long time. Note: ClientCollateralPerEpoch may not be needed and removed pending future confirmation. There will be a Minimum value for both client and provider deal collateral.

DealQueryParams

DealSettlementSummary

DealSpaces

DealState

GetBalanceParams

GetBalanceReturn

GetDealActivationReturn

GetDealClientCollateralReturn

GetDealClientReturn

GetDealDataCommitmentReturn

GetDealLabelReturn

GetDealProviderCollateralReturn

GetDealProviderReturn

GetDealSectorReturn

GetDealTermReturn

GetDealTotalPriceReturn

GetDealVerifiedReturn

MarketNotifyDealParams

OnMinerSectorsTerminateParams

PublishStorageDealsParams

PublishStorageDealsReturn

SectorDataSpec

SectorDealActivation

SectorDealIDs

IDs of deals associated with a single sector.

SectorDeals

SettleDealPaymentsParams

SettleDealPaymentsReturn

State

Market actor state

VerifyDealsForActivationParams

VerifyDealsForActivationReturn

WithdrawBalanceParams

WithdrawBalanceReturn

Enums

Label

LoadDealState

Method

Market actor methods available

Reason

Constants

EX_DEAL_EXPIRED

EX_DEAL_NOT_ACTIVATED

MARKET_NOTIFY_DEAL_METHOD

NO_ALLOCATION_ID

PENDING_ALLOCATIONS_CONFIG

PROPOSALS_AMT_BITWIDTH

PROVIDER_SECTORS_CONFIG

SECTOR_DEALS_CONFIG

STATES_AMT_BITWIDTH

Functions

deal_get_payment_remaining

find_deal_state

find_proposal

get_proposal

has_pending_deal

is_piece_cid

Cid prefix for piece Cids

load_provider_sector_deals

Type Aliases

DealArray

A specialization of a array to deals.

DealMetaArray

A specialization of a array to deals.

GetDealActivationParams

GetDealClientCollateralParams

GetDealClientParams

GetDealDataCommitmentParams

GetDealLabelParams

GetDealProviderCollateralParams

GetDealProviderParams

GetDealSectorParams

GetDealTermParams

GetDealTotalPriceParams

GetDealVerifiedParams

PendingDealAllocationsMap

ProviderSectorsMap

SectorDealsMap