Module fil_actor_market_state::v9

Modules

balance_table

Structs

ActivateDealsParams

ActivateDealsResult

ClientDealProposal

ClientDealProposal is a DealProposal signed by a client

ComputeDataCommitmentParams

ComputeDataCommitmentParamsRef

ComputeDataCommitmentReturn

DealProposal

Note: Deal Collateral is only released and returned to clients and miners when the storage deal stops counting towards power. In the current iteration, it will be released when the sector containing the storage deals expires, even though some storage deals can expire earlier than the sector does. Collaterals are denominated in PerEpoch to incur a cost for self dealing or minimal deals that last for a long time. Note: ClientCollateralPerEpoch may not be needed and removed pending future confirmation. There will be a Minimum value for both client and provider deal collateral.

DealSpaces

DealState

OnMinerSectorsTerminateParams

OnMinerSectorsTerminateParamsRef

PublishStorageDealsParams

PublishStorageDealsReturn

SectorDataSpec

SectorDealData

SectorDeals

State

Market actor state

VerifiedDealInfo

VerifyDealsForActivationParams

VerifyDealsForActivationParamsRef

VerifyDealsForActivationReturn

WithdrawBalanceParams

WithdrawBalanceReturn

Enums

Label

Method

Market actor methods available

Constants

PROPOSALS_AMT_BITWIDTH

STATES_AMT_BITWIDTH

Type Aliases

DealArray

A specialization of a array to deals.

DealMetaArray

A specialization of a array to deals.